Udacity OpenVAS Scan
This is the Udacity OpenVAS scan.  Unable to get the Virtual Machines to talk to each other.  Could ping from VM to Localhost, but not the other way around.  The two initial Docker images installed and was able to scan one Docker image with the OpenVAS Docker image.  Again, hours of my life spent troubleshooting network connectivity issues in Virtual Machines, which is outside the scope of the assignment.  Hopefully this early image will show you that as long as the technology/instructions line up and things work as advertised, I can use the tools and read the output to determine the vulnerabilities that exist and analyze the potential risk inherent to my systems/networks.  In the real-world I have others on my team, such as the Network and Linux Administrators that would be able to assist me in troubleshooting connectivity issues.