Update backdoor_details.txt
Made changes to include the output of the following PS commands to obtain security information:
          ps -eo euser,ruser,suser,fuser,f,comm,label
          ps axZ
          ps -eM 
   
All of this information can be analyzed to determine malicious activity on the network.  By fetching the list of all processes running and also keeping an eye on what commands are used to run the processes, checking all of the ports that are actively listening, and analyzing the ports for any unusual port numbers we would be able to triage for useful information that would help us to build a complete picture of any backdoor activity details.

I am still unable to install and use the ossec web UI on the provided Ubuntu Virtual-Box image so I cannot appropriately analyze the ossec logs.  In the real-world I would use a supported HID that I could actually install and use in my program's environment.